+++ NEWS.md
@@ -1,3 +1,10 @@
+## metainc, version 0.2-2 (2025-mm-dd)
+
+- R function heatplot():
+  - converted the function into a generic function
+  - return ggplot object(s) instead of NULL
+
+
 ## metainc, version 0.2-1 (2025-03-11)
 
 - Small revisions due to the publication of